fix(no-entire-object-hook-deps): Improve object usage detection and add specific test case
- Enhance object usage detection in hook dependencies
- Add more robust filtering for array and object path detection
- Remove separate bug reproduction test file
- Add comprehensive test case for theme object usage
- Improve error reporting for specific object field dependencies
